add domain classes

+++ app/src/main/java/com/shpzdsh/shopinglist/domain/ShopItem.kt
@@ -3,6 +3,6 @@ package com.shpzdsh.shopinglist.domain
 data class ShopItem(
     val id: Int,
     val name: String,
-    val count: Int,
+    val count: Double,
     val enabled: Boolean
 )
